Add case-insensitive option for tag names.
This commit was adapted from an unmerged pull request [#263](https://github.com/alex/django-taggit/pull/263) done by [qris](https://github.com/qrisa). The following text is his original commit message:

Add case-insensitive option for tag names.

I don't think it makes sense to have tags which differ only in case. It's not even possible on MySQL because it violates the
unique constraint on tag name:

    IntegrityError: (1062, "Duplicate entry 'LONDON' for key 'taggit_tag_name_6b48d50f_uniq'")

This PR adds a new setting, `TAGGIT_CASE_INSENSITIVE`, which causes existing tags to be looked up case insensitively. New
tags are added with the supplied case, which will be used from then on. The default behaviour is the same as before without
this setting.

This meets nookiepl's request on #70, and my comment on #9.